requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy>=1.26.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
supabase==1.0.3
python-dotenv==1.0.0
requests>=2.31.0
